@@ -30,16 +30,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.9",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.0.5",
-    "@angular/common": "^17.0.9",
-    "@angular/compiler": "^17.0.9",
-    "@angular/core": "^17.0.9",
-    "@angular/forms": "^17.0.9",
-    "@angular/platform-browser": "^17.0.9",
-    "@angular/platform-browser-dynamic": "^17.0.9",
-    "@angular/router": "^17.0.9",
-    "@angular/service-worker": "^17.0.9",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
+    "@angular/service-worker": "^18.2.13",
     "@fortawesome/angular-fontawesome": "^0.14.1",
     "@fortawesome/fontawesome-svg-core": "^6.5.1",
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
@@ -62,9 +62,9 @@
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.10",
-    "@angular/cli": "^17.0.10",
-    "@angular/compiler-cli": "^17.0.9",
+    "@angular-devkit/build-angular": "^18.2.14",
+    "@angular/cli": "^18.2.14",
+    "@angular/compiler-cli": "^18.2.13",
     "@types/d3": "^7.4.3",
     "@types/jasmine": "~4.0.3",
     "@types/lodash-es": "^4.17.12",
@@ -81,7 +81,7 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^3.2.2",
     "pretty-quick": "^3.1.3",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.4"
   },
   "husky": {
     "hooks": {
